requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.2

# Database
psycopg2-binary==2.9.9

# Image Processing & Computer Vision
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
tensorflow==2.14.0
torch==2.1.1
torchvision==0.16.1
statsmodels==0.14.0
prophet==1.1.5

# Async & Caching
celery==5.3.4
redis==5.0.1

# Excel & CSV Processing
openpyxl==3.11.0
xlrd==2.0.1
pandas==2.1.3

# API Clients
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.0
marshmallow==3.20.1

# Logging & Monitoring
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# YOLO for Computer Vision
ultralytics==8.0.208

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Error Tracking
sentry-sdk==1.38.0

# API Rate Limiting
djangorestframework-ratelimit==1.0.1

# Pagination & Serialization
django-rest-nested-routers==0.92.5
